refactor(utils): split cdh_utils.py into a package
The single 1643-line utils/cdh_utils.py module is replaced by a
utils/cdh_utils/ package whose private submodules group helpers by
concern:

* _common      — shared "logger" and "F" TypeVar.
* _dates       — Pega date-time parsing (from/to_prpc_date_time,
                 parse_pega_date_time_formats, _get_start_end_date_args).
* _namespacing — _capitalize and default_predictor_categorization.
* _polars      — query/sample/schema/overlap/weighted-average helpers
                 (_apply_query, _combine_queries, _polars_capitalize,
                 _extract_keys, _apply_schema_types, lazy_sample,
                 weighted_average_polars, weighted_performance_polars,
                 overlap_matrix, overlap_lists_polars,
                 is_valid_polars_duration, POLARS_DURATION_PATTERN).
* _metrics     — AUC / lift / log-odds / gains / feature importance.
* _io          — file, working-dir, logger setup, version-check helpers.
* _misc        — safe_flatten_list, legend_color_order.

The package __init__ re-exports the exact same public surface the flat
module previously exposed (verified by diffing the dir() output before
and after) so all existing imports such as
"from pdstools.utils.cdh_utils import safe_range_auc" keep working
unchanged. Function bodies are untouched — this is a pure
code-organisation change.
